Rename App components.
