+++ example-wasm/src/App.tsx
@@ -1,33 +1,39 @@
-import { Navigate, Outlet } from 'react-router-dom';
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
-import { Task } from './models/Task';
-import { PageLayout } from './components/PageLayout';
+import { AuthenticatedApp } from './AuthenticatedApp';
+import { ErrorPage } from './pages/ErrorPage';
+import { LoginPage } from './pages/LoginPage';
+import { TaskPage } from './pages/TaskPage';
+import config from './atlas-app-services/config.json';
+import styles from './styles/AppWrapper.module.css';
 
-const { RealmProvider, UserProvider, useApp } = await import('@realm/react');
+const { AppProvider } = await import('@realm/react');
 
-export function App() {
-  const atlasApp = useApp();
-  if (!atlasApp.currentUser) {
-    return <Navigate to='/' />
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <LoginPage />,
+    errorElement: <ErrorPage />,
+  },
+  {
+    element: <AuthenticatedApp />,
+    children: [
+      {
+        path: 'tasks',
+        element: <TaskPage />
+      }
+    ]
   }
+]);
 
+function App() {
   return (
-    <UserProvider>
-      <RealmProvider
-        schema={[Task]}
-        sync={{
-          flexible: true,
-          initialSubscriptions: {
-            update: ((mutableSubs, realm) => {
-              mutableSubs.add(realm.objects(Task), { name: 'allTasks' });
-            }),
-          },
-        }}
-      >
-        <PageLayout>
-          <Outlet />
-        </PageLayout>
-      </RealmProvider>
-    </UserProvider>
+    <div className={styles.container}>
+      <AppProvider id={config.ATLAS_APP_ID}>
+        <RouterProvider router={router} />
+      </AppProvider>
+    </div>
   );
 }
+
+export default App;
