+++ example-wasm/src/AuthenticatedApp.tsx
@@ -1,0 +1,33 @@
+import { Navigate, Outlet } from 'react-router-dom';
+
+import { Task } from './models/Task';
+import { PageLayout } from './components/PageLayout';
+
+const { RealmProvider, UserProvider, useApp } = await import('@realm/react');
+
+export function AuthenticatedApp() {
+  const atlasApp = useApp();
+  if (!atlasApp.currentUser) {
+    return <Navigate to='/' />
+  }
+
+  return (
+    <UserProvider>
+      <RealmProvider
+        schema={[Task]}
+        sync={{
+          flexible: true,
+          initialSubscriptions: {
+            update: ((mutableSubs, realm) => {
+              mutableSubs.add(realm.objects(Task), { name: 'allTasks' });
+            }),
+          },
+        }}
+      >
+        <PageLayout>
+          <Outlet />
+        </PageLayout>
+      </RealmProvider>
+    </UserProvider>
+  );
+}
