+++ example-wasm/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import AppWrapper from './AppWrapper';
+import App from './App';
 import reportWebVitals from './reportWebVitals';
 import './styles/global.css';
 
@@ -10,7 +10,7 @@ const root = ReactDOM.createRoot(
 );
 root.render(
   <React.StrictMode>
-    <AppWrapper />
+    <App />
   </React.StrictMode>
 );
